requirements.txt
aiocache==0.12.2
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
Faker==24.0.0
fastapi==0.110.0
fastapi-cli==0.0.4
graphql-core==3.2.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
neo4j==5.19.0
neomodel==5.3.1
orjson==3.10.5
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.18.0
pytest==8.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
setuptools==69.0.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
strawberry-graphql==0.235.0
typer==0.12.3
typing_extensions==4.10.0
ujson==5.10.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.42.0
